fix ASTForm type issues for build
